python-graph-topology/requirements.txt
nuitka==0.6.8.1
mysql==0.0.2
graphviz==0.8.4
pydot==1.4.1
pandas==0.24.2
networkx==2.4
pyyed==1.3
matplotlib==3.0.2
pyvis==0.1.8.2
geopy==2.0.0
folium==0.11.0
pygraphviz==1.5
openpyxl==2.6.4